% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatRater.R
\name{generate_ratings}
\alias{generate_ratings}
\title{Generate ratings for a single stim using ChatGPT}
\usage{
generate_ratings(
  model = "gpt-3.5-turbo",
  stim,
  prompt,
  question,
  temp = 0,
  n_iterations = 30,
  api_key,
  debug = FALSE
)
}
\arguments{
\item{model}{A character string specifying the ChatGPT model to use.}

\item{stim}{A character string representing the stim (e.g., an idiom).}

\item{prompt}{A character string providing context or an identity for ChatGPT (e.g., "You are a native English speaker.").}

\item{question}{A character string that provides instructions for ChatGPT.}

\item{temp}{A numeric value specifying the temperature for the API call.}

\item{n_iterations}{An integer indicating the number of times to query ChatGPT for the stim.}

\item{api_key}{Your OpenAI API key.}

\item{debug}{Logical flag indicating whether to display debugging information. Default is FALSE.}
}
\value{
A data frame containing the stim, rating, and iteration number for each API call.
}
\description{
Generate ratings for a single stim using ChatGPT
}
