% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatRater.R
\name{generate_ratings_for_all}
\alias{generate_ratings_for_all}
\title{Generate ratings for all stims using ChatGPT}
\usage{
generate_ratings_for_all(
  model = "gpt-3.5-turbo",
  stim_list,
  prompt,
  question,
  temp = 0,
  n_iterations = 30,
  api_key,
  debug = FALSE
)
}
\arguments{
\item{model}{A character string specifying the ChatGPT model to use.}

\item{stim_list}{A character vector of stims (e.g., idioms) for which ratings will be generated.}

\item{prompt}{A character string providing context or an identity for ChatGPT (e.g., "You are a native English speaker.").}

\item{question}{A character string that provides instructions for ChatGPT.}

\item{temp}{A numeric value specifying the temperature for the API call.}

\item{n_iterations}{An integer indicating the number of times to query ChatGPT for each stim.}

\item{api_key}{Your OpenAI API key.}

\item{debug}{Logical flag indicating whether to display debugging information. Default is FALSE.}
}
\value{
A data frame containing the stim, rating, and iteration number for each API call.
}
\description{
This function iterates over a list of stims (e.g., idioms) and generates ratings for each
by calling the \code{generate_ratings} function. It aggregates all results into a single data frame.
}
