% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatRater.R
\name{generate_ratings_for_all}
\alias{generate_ratings_for_all}
\title{Batch Rating Generator}
\usage{
generate_ratings_for_all(model = "gpt-4-turbo", stim_list, ...)
}
\arguments{
\item{model}{LLM model name (default: 'gpt-4-turbo')}

\item{stim_list}{List of stimuli to process}
}
\description{
Process multiple stimuli in sequence
}
